requirements.txt
requests
asgiref==3.9.2
beautifulsoup4==4.14.2
Django==4.2.25
django-bootstrap-v5==1.0.11
python-decouple==3.8
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
